Amazon Bedrock Agentcore/agent/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
bedrock-agentcore-starter-toolkit